// src/types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
// ...
      consultations: {
        Row: {
          address: string | null
          base_fee: number
          clinical_notes: string | null
          created_at: string
          date: string
          id: string
          images: string[] | null
          patient_id: string
          profile_id: string
          status: string
          type: string
          weight_kg: number | null
        }
        Insert: {
          address?: string | null
          base_fee?: number
          clinical_notes?: string | null
          created_at?: string
          date?: string
          id?: string
          images?: string[] | null
          patient_id: string
          profile_id: string
          status?: string
          type: string
          weight_kg?: number | null
        }
        Update: {
          address?: string | null
          base_fee?: number
          clinical_notes?: string | null
          created_at?: string
          date?: string
          id?: string
          images?: string[] | null
          patient_id?: string
          profile_id?: string
          status?: string
          type?: string
          weight_kg?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "consultations_patient_id_fkey"
            columns: ["patient_id"]
            isOneToOne: false
            referencedRelation: "patients"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "consultations_profile_id_fkey"
            columns: ["profile_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      decrement_inventory: {
        Args: { amount: number; item_id: string }
        Returns: undefined
      }
    }
    Enums: {
      invoice_status: "pending" | "paid" | "cancelled"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}